deposit backend done

+++ Backend/RLNCustomerApplication/src/main/java/com/rln/model/Customer.java
@@ -58,7 +58,7 @@ public class Customer {
 	@Column(nullable = false)
 	private String accountType;
 	@Column(nullable = false)
-	private long balance = (long) 25000;
+	private double balance = 25000.00;
 	@Column(nullable = false)
 	private boolean verificationStatus = false;
 	@Column(nullable = false, updatable = false)
